requirements.txt
langchain
openai
load_dotenv
langserve[all]
qdrant-client
tiktoken
fastapi